SQL hikari add & reformat & idek

+++ common-sql/src/main/java/com/oop/datamodule/commonsql/util/TableEditor.java
@@ -2,49 +2,53 @@ package com.oop.datamodule.commonsql.util;
 
 import com.oop.datamodule.api.util.DataPair;
 import com.oop.datamodule.commonsql.database.SQLDatabase;
-import lombok.RequiredArgsConstructor;
-
 import java.util.ArrayList;
 import java.util.List;
+import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public class TableEditor {
-    private final String table;
-    private final List<DataPair<String, String>> addColumns = new ArrayList<>();
-    private final List<String> dropsColumns = new ArrayList<>();
-    private final List<DataPair<String, String>> renamedColumns = new ArrayList<>();
+  private final String table;
+  private final List<DataPair<String, String>> addColumns = new ArrayList<>();
+  private final List<String> dropsColumns = new ArrayList<>();
+  private final List<DataPair<String, String>> renamedColumns = new ArrayList<>();
 
-    public TableEditor addColumn(String name, String type) {
-        addColumns.add(new DataPair<>(name, type));
-        return this;
+  public TableEditor addColumn(String name, String type) {
+    addColumns.add(new DataPair<>(name, type));
+    return this;
+  }
+
+  public TableEditor addDropColumn(String column) {
+    dropsColumns.add(column);
+    return this;
+  }
+
+  public TableEditor renameColumn(String oldColumnName, String newColumnName) {
+    renamedColumns.add(new DataPair<>(oldColumnName, newColumnName));
+    return this;
+  }
+
+  public void edit(SQLDatabase database) {
+    if (!database.getTables().contains(table)) return;
+
+    // Add columns
+    for (DataPair<String, String> column : addColumns) {
+      List<String> columns = database.getColumns(table);
+      if (!columns.contains(column.getKey()))
+        database.execute(
+            "ALTER TABLE "
+                + table
+                + " ADD "
+                + SqlUtil.escapeColumn(column.getKey(), database)
+                + " "
+                + column.getValue());
     }
 
-    public TableEditor addDropColumn(String column) {
-        dropsColumns.add(column);
-        return this;
-    }
+    // Rename columns
+    if (!renamedColumns.isEmpty())
+      database.renameColumn(table, renamedColumns.toArray(new DataPair[0]));
 
-    public TableEditor renameColumn(String oldColumnName, String newColumnName) {
-        renamedColumns.add(new DataPair<>(oldColumnName, newColumnName));
-        return this;
-    }
-
-    public void edit(SQLDatabase database) {
-        if (!database.getTables().contains(table)) return;
-
-        // Add columns
-        for (DataPair<String, String> column : addColumns) {
-            List<String> columns = database.getColumns(table);
-            if (!columns.contains(column.getKey()))
-                database.execute("ALTER TABLE " + table + " ADD " + SqlUtil.escapeColumn(column.getKey(), database) + " " + column.getValue());
-        }
-
-        // Rename columns
-        if (!renamedColumns.isEmpty())
-            database.renameColumn(table, renamedColumns.toArray(new DataPair[0]));
-
-        // Drop columns
-        if (!dropsColumns.isEmpty())
-            database.dropColumn(table, dropsColumns.toArray(new String[0]));
-    }
+    // Drop columns
+    if (!dropsColumns.isEmpty()) database.dropColumn(table, dropsColumns.toArray(new String[0]));
+  }
 }
